game object position en float maintenant

+++ Arkanoid/SDL2_OO/sdl_o_rectangle.cpp
@@ -1,12 +1,12 @@
 #include "sdl_o_rectangle.h"
 
-Sdl_o_rectangle::Sdl_o_rectangle(int x1, int y1, int w1, int h1)
+Sdl_o_rectangle::Sdl_o_rectangle(float x1, float y1, float w1, float h1)
 {
   m_x = x1;
   m_y = y1;
   m_width = w1;
   m_height = h1;
-  rect = { m_x, m_y, m_width, m_height };
+  rect = { static_cast<int>(m_x), static_cast<int>(m_y), static_cast<int>(m_width), static_cast<int>(m_height) };
 };
 
 Sdl_o_rectangle::Sdl_o_rectangle()
@@ -15,22 +15,22 @@ Sdl_o_rectangle::Sdl_o_rectangle()
   m_y = 0;
   m_width = 0;
   m_height = 0;
-  rect = { m_x, m_y, m_width, m_height };
+  rect = { static_cast<int>(m_x), static_cast<int>(m_y), static_cast<int>(m_width), static_cast<int>(m_height) };
 }
 
 SDL_Rect* Sdl_o_rectangle::getRect()
 {
-  rect = { m_x, m_y, m_width, m_height };//update rect struct with attribute before returning rect
+  rect = { static_cast<int>(m_x), static_cast<int>(m_y), static_cast<int>(m_width), static_cast<int>(m_height) };//update rect struct with attribute before returning rect
   return &rect;
 }
 
-void Sdl_o_rectangle::setX(int x)
+void Sdl_o_rectangle::setX(float x)
 {
-  m_x = x;
+  m_x = static_cast<int>(x);
 }
 
 
-void Sdl_o_rectangle::setY(int y)
+void Sdl_o_rectangle::setY(float y)
 {
-  m_y = y;
+  m_y = static_cast<int>(y);
 }
