+++ Arkanoid/SDL2_OO/sdl_o_rectangle.h
@@ -5,19 +5,19 @@
 class Sdl_o_rectangle
 {
   public:
-    int m_x = 0;
-    int m_y = 0;
-    int m_width = 0;
-    int m_height = 0;
+    float m_x = 0;
+    float m_y = 0;
+    float m_width = 0;
+    float m_height = 0;
   private:
     SDL_Rect rect;
 
  public:
    Sdl_o_rectangle();
-   Sdl_o_rectangle(int x1, int y1, int w1, int h1);
+   Sdl_o_rectangle(float x1, float y1, float w1, float h1);
    SDL_Rect *getRect();
-   void setX(int x);
-   void setY(int y);
+   void setX(float x);
+   void setY(float y);
 };
 
 #endif
